Move validation tree view to components
